+++ sphinx/builders/texinfo.py
@@ -186,8 +186,6 @@ class TexinfoBuilder(Builder):
                         nodes.Text(')'),
                     ))
                     break
-            else:
-                pass
             pendingnode.replace_self(newnodes)
         return largetree