+++ sphinx/ext/imgmath.py
@@ -366,7 +366,7 @@ def html_visit_math(self: HTML5Translator, node: nodes.math) -> None:
 
 
 def html_visit_displaymath(self: HTML5Translator, node: nodes.math_block) -> None:
-    if node['nowrap']:
+    if node['no-wrap'] or node['nowrap']:
         latex = node.astext()
     else:
         latex = wrap_displaymath(node.astext(), None, False)