+++ sphinx/util/typing.py
@@ -2,13 +2,22 @@
 
 from __future__ import annotations
 
+import contextvars
+import ctypes
 import dataclasses
+import io
+import json
+import lzma
+import multiprocessing
+import pathlib
+import pickle  # NoQA: S403
+import struct
 import sys
 import types
 import typing
+import weakref
+import zipfile
 from collections.abc import Callable, Sequence
-from contextvars import Context, ContextVar, Token
-from struct import Struct
 from typing import TYPE_CHECKING
 
 from docutils import nodes
@@ -40,10 +49,43 @@ logger = logging.getLogger(__name__)
 
 # classes that have an incorrect .__module__ attribute
 _INVALID_BUILTIN_CLASSES: Final[Mapping[object, str]] = {
-    Context: 'contextvars.Context',  # Context.__module__ == '_contextvars'
-    ContextVar: 'contextvars.ContextVar',  # ContextVar.__module__ == '_contextvars'
-    Token: 'contextvars.Token',  # Token.__module__ == '_contextvars'
-    Struct: 'struct.Struct',  # Struct.__module__ == '_struct'
+    # types in 'contextvars' with <type>.__module__ == '_contextvars':
+    contextvars.Context: 'contextvars.Context',
+    contextvars.ContextVar: 'contextvars.ContextVar',
+    contextvars.Token: 'contextvars.Token',
+    # types in 'ctypes' with <type>.__module__ == '_ctypes':
+    ctypes.Array: 'ctypes.Array',
+    ctypes.Structure: 'ctypes.Structure',
+    ctypes.Union: 'ctypes.Union',
+    # types in 'io' with <type>.__module__ == '_io':
+    io.FileIO: 'io.FileIO',
+    io.BytesIO: 'io.BytesIO',
+    io.StringIO: 'io.StringIO',
+    io.BufferedReader: 'io.BufferedReader',
+    io.BufferedWriter: 'io.BufferedWriter',
+    io.BufferedRWPair: 'io.BufferedRWPair',
+    io.BufferedRandom: 'io.BufferedRandom',
+    io.TextIOWrapper: 'io.TextIOWrapper',
+    # types in 'json' with <type>.__module__ == 'json.{decoder,encoder}':
+    json.JSONDecoder: 'json.JSONDecoder',
+    json.JSONEncoder: 'json.JSONEncoder',
+    # types in 'lzma' with <type>.__module__ == '_lzma':
+    lzma.LZMACompressor: 'lzma.LZMACompressor',
+    lzma.LZMADecompressor: 'lzma.LZMADecompressor',
+    # types in 'multiprocessing' with <type>.__module__ == 'multiprocessing.context':
+    multiprocessing.Process: 'multiprocessing.Process',
+    # types in 'pathlib' with <type>.__module__ == 'pathlib._local':
+    pathlib.Path: 'pathlib.Path',
+    pathlib.PosixPath: 'pathlib.PosixPath',
+    pathlib.PurePath: 'pathlib.PurePath',
+    pathlib.PurePosixPath: 'pathlib.PurePosixPath',
+    pathlib.PureWindowsPath: 'pathlib.PureWindowsPath',
+    pathlib.WindowsPath: 'pathlib.WindowsPath',
+    # types in 'pickle' with <type>.__module__ == 'pickle':
+    pickle.Pickler: 'pickle.Pickler',
+    pickle.Unpickler: 'pickle.Unpickler',  # NoQA: S301
+    # types in 'struct' with <type>.__module__ == '_struct':
+    struct.Struct: 'struct.Struct',
     # types in 'types' with <type>.__module__ == 'builtins':
     types.AsyncGeneratorType: 'types.AsyncGeneratorType',
     types.BuiltinFunctionType: 'types.BuiltinFunctionType',
@@ -52,6 +94,7 @@ _INVALID_BUILTIN_CLASSES: Final[Mapping[object, str]] = {
     types.ClassMethodDescriptorType: 'types.ClassMethodDescriptorType',
     types.CodeType: 'types.CodeType',
     types.CoroutineType: 'types.CoroutineType',
+    types.EllipsisType: 'types.EllipsisType',
     types.FrameType: 'types.FrameType',
     types.FunctionType: 'types.FunctionType',
     types.GeneratorType: 'types.GeneratorType',
@@ -63,8 +106,15 @@ _INVALID_BUILTIN_CLASSES: Final[Mapping[object, str]] = {
     types.MethodType: 'types.MethodType',
     types.MethodWrapperType: 'types.MethodWrapperType',
     types.ModuleType: 'types.ModuleType',
+    types.NoneType: 'types.NoneType',
+    types.NotImplementedType: 'types.NotImplementedType',
     types.TracebackType: 'types.TracebackType',
     types.WrapperDescriptorType: 'types.WrapperDescriptorType',
+    # types in 'weakref' with <type>.__module__ == '_weakrefset':
+    weakref.WeakSet: 'weakref.WeakSet',
+    # types in 'zipfile' with <type>.__module__ == 'zipfile._path':
+    zipfile.Path: 'zipfile.Path',
+    zipfile.CompleteDirs: 'zipfile.CompleteDirs',
 }